+++ src/applications/IsIdentifiableReviewer/Out/RowUpdater.cs
@@ -1,25 +1,50 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Text;
 using FAnsi.Discovery;
 using Microservices.IsIdentifiable.Reporting;
+using Microservices.IsIdentifiable.Rules;
 
 namespace IsIdentifiableReviewer.Out
 {
-    public class RowUpdater
+    public class RowUpdater : OutBase
     {
+        public const string DefaultFileName = "RedList.yaml";
+
+        /// <summary>
+        /// Set to true to only output updates to redlist instead of trying to update the database.
+        /// This is useful if you want to  run in manual mode to process everything then run unattended
+        /// for the updates.
+        /// </summary>
+        public bool RulesOnly { get; set; }
+
         Dictionary<DiscoveredTable,DiscoveredColumn> _primaryKeys = new Dictionary<DiscoveredTable, DiscoveredColumn>();
-
-        List<Failure> _committed = new List<Failure>();
-
-        public void Update(Target target, Failure failure)
+        
+        public RowUpdater(FileInfo rulesFile) : base(rulesFile)
         {
-            Update(target.Discover(),failure);
         }
 
-        public void Update(DiscoveredServer server, Failure failure)
+        public RowUpdater() : this(new FileInfo(DefaultFileName))
         {
+        }
+
+        public void Update(Target target, Failure failure, bool addRule)
+        {
+            Update(target.Discover(),failure,addRule);
+        }
+
+        public void Update(DiscoveredServer server, Failure failure, bool addRule)
+        {
+            //add the update rule to the redlist
+            if(addRule)
+                Add(failure,RuleAction.Report);
+
+            //if we are running in rules only mode we don't need to also update the database
+            if(RulesOnly)
+                return;
+
             var syntax = server.GetQuerySyntaxHelper();
 
             //the fully specified name e.g. [mydb]..[mytbl]
@@ -61,28 +86,26 @@ namespace IsIdentifiableReviewer.Out
                     cmd.ExecuteNonQuery();
                 }   
             }
-            
-            _committed.Add(failure);
         }
 
         /// <summary>
         /// Handler for loading <paramref name="failure"/>.  If the user previously made an update decision an
         /// update will transparently happen for this record and false is returned.
         /// </summary>
-        /// <param name="currentTarget"></param>
+        /// <param name="server"></param>
         /// <param name="failure"></param>
         /// <returns>True if <paramref name="failure"/> is novel and not seen before</returns>
-        public bool OnLoad(Target currentTarget,Failure failure)
+        public bool OnLoad(DiscoveredServer server,Failure failure)
         {
             //we have bigger problems than if this is novel!
-            if (currentTarget == null)
+            if (server == null)
                 return true;
 
             //if we have seen this before
-            if (_committed.Any(c => c.HaveSameProblem(failure)))
+            if (IsCoveredByExistingRule(failure))
             {
                 //since user has issued an update for this exact problem before we can update this one too
-                Update(currentTarget,failure);
+                Update(server,failure,false);
 
                 //and return false to indicate that it is not a novel issue
                 return false;
@@ -90,5 +113,6 @@ namespace IsIdentifiableReviewer.Out
 
             return true;
         }
+
     }
 }