+++ src/microservices/Microservices.IsIdentifiable/Service/IsIdentifiableQueueConsumer.cs
@@ -15,12 +15,14 @@ namespace Microservices.IsIdentifiable.Service
     {
         private readonly IProducerModel _producer;
         private readonly string _fileSystemRoot;
+        private readonly string _extractionRoot;
         private readonly IClassifier _classifier;
 
-        public IsIdentifiableQueueConsumer(IProducerModel producer, string fileSystemRoot, IClassifier classifier)
+        public IsIdentifiableQueueConsumer(IProducerModel producer, string fileSystemRoot, string extractionRoot, IClassifier classifier)
         {
             _producer = producer;
             _fileSystemRoot = fileSystemRoot;
+            _extractionRoot = extractionRoot;
             _classifier = classifier;
         }
 
@@ -36,12 +38,11 @@ namespace Microservices.IsIdentifiable.Service
 
             // The path is taken from the message, however maybe it should be FileSystemOptions|ExtractRoot in default.yaml
             // If the filename has a rooted path then the ExtractionDirectory is ignored by Path.Combine
-            // TODO(rkm 2020-02-04) Check that CTP doesn't output rooted paths / assert here
-            var toProcess = new FileInfo( Path.Combine(message.ExtractionDirectory, message.AnonymisedFileName) );
+            var toProcess = new FileInfo( Path.Combine(_extractionRoot, message.ExtractionDirectory, message.AnonymisedFileName) );
 
             if(!toProcess.Exists)
                 //  XXX  this causes a fatal error and the whole service terminates
-                throw new FileNotFoundException();
+                throw new FileNotFoundException("IsIdentifiable service cannot find file "+toProcess.FullName);
 
             var result = _classifier.Classify(toProcess);